making tag component folder
